\name{GENESISasian-package}
\alias{GENESISasian-package}
\alias{GENESISasian}
\docType{package}
\title{
\packageTitle{GENESISasian}
}
\description{
\packageDescription{GENESISasian}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GENESISasian}
\packageIndices{GENESISasian}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GENESISasian}

Maintainer: \packageMaintainer{GENESISasian}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
